requirements.txt
openai
streamlit
python-dotenv
sentence-transformers 
pinecone
pyreadline3
langchain
langchain_openai